Move LatestCharmInfo() to its own file.
