+++ charmstore/client.go
@@ -6,7 +6,6 @@ package charmstore
 import (
 	"io"
 	"io/ioutil"
-	"time"
 
 	"github.com/juju/errors"
 	"github.com/juju/loggo"
@@ -174,48 +173,6 @@ func (c *Client) LatestRevisions(cURLs []*charm.URL) ([]charmrepo.CharmRevision,
 }
 
 // TODO(ericsnow) Add an "AsRepo() charmrepo.Interface" method.
-
-// LatestCharmInfo returns the most up-to-date information about each
-// of the identified charms at their latest revision. The revisions in
-// the provided URLs are ignored.
-func LatestCharmInfo(client BaseClient, cURLs []*charm.URL) ([]CharmInfoResult, error) {
-	now := time.Now().UTC()
-
-	// Do a bulk call to get the revision info for all charms.
-	logger.Infof("retrieving revision information for %d charms", len(cURLs))
-	revResults, err := client.LatestRevisions(cURLs)
-	if err != nil {
-		err = errors.Annotate(err, "while getting latest charm revision info")
-		logger.Infof(err.Error())
-		return nil, err
-	}
-
-	// Do a bulk call to get the latest info for each charm's resources.
-	// TODO(ericsnow) Only do this for charms that *have* resources.
-	chResources, err := client.ListResources(cURLs)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-
-	// Extract the results.
-	var results []CharmInfoResult
-	for i, cURL := range cURLs {
-		revResult := revResults[i]
-		resources := chResources[i]
-
-		var result CharmInfoResult
-		result.OriginalURL = cURL
-		result.Timestamp = now
-		if revResult.Err != nil {
-			result.Error = errors.Trace(revResult.Err)
-		} else {
-			result.LatestRevision = revResult.Revision
-			result.LatestResources = resources
-		}
-		results = append(results, result)
-	}
-	return results, nil
-}
 
 // TODO(ericsnow) Remove the fake once the charm store adds support.
